Updated Home View Implementation
