+++ Showcase-iOS/Dashboard/Cell/ShowcaseAppCollectionViewCell.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import Kingfisher
 
 class ShowcaseAppCollectionViewCell: UICollectionViewCell {
     @IBOutlet private weak var imageView: UIImageView!
@@ -31,7 +32,11 @@ class ShowcaseAppCollectionViewCell: UICollectionViewCell {
         let imageFetcher = ImageFetcher(from: firStorage)
         imageFetcher.fetchImage(imageUrl) { (url, error) in
             if error == nil {
-                
+                guard let imageURL = url else {
+                    return
+                }
+                let resource = ImageResource(downloadURL: imageURL, cacheKey: imageUrl)
+                self.imageView.kf.setImage(with: resource, placeholder: nil, options: nil, progressBlock: nil, completionHandler: nil)
             }
         }
     }
