--- Showcase-iOS/Dashboard/Protocol/InjectView.swift
@@ -1,13 +1,0 @@
-//
-//  InjectView.swift
-//  Showcase-iOS
-//
-//  Created by Edward Mtshweni on 2018/05/14.
-//  Copyright © 2018 DVT. All rights reserved.
-//
-
-import Foundation
-
-protocol InjectView {
-    func inject(with storyBoard: AppStoryBoard)
-}
